Refactoring: Made project structure consistent.
