+++ packages/esm-patient-list-app/src/add-patient/add-patient.scss
@@ -26,7 +26,7 @@
   overflow: hidden;
 }
 
-.paginationContainer { 
+.paginationContainer {
   display: flex;
   justify-content: space-between;
   align-items: center;
