+++ packages/esm-patient-list-app/src/api/index.ts
@@ -1,5 +1,5 @@
+export * from './types';
 export * from './api-remote';
 export * from './api-local';
 export * from './api';
 export * from './queries';
-export * from './types';
