+++ packages/esm-patient-list-app/src/patient-list-link.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { ConfigurableLink } from '@openmrs/esm-framework';
 
-export default function () {
+export default function PatientListLink() {
   const { t } = useTranslation();
   return (
     <ConfigurableLink to="${openmrsSpaBase}/patient-list">
