+++ packages/esm-patient-list-app/src/patient-list-list/patient-list-list.scss
