+++ packages/esm-patient-list-app/src/patient-table/patient-table.component.scss
